Change console widgets imports after moving.
